requirements.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dominate==2.9.1
flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.1.0
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.7.0
flask-sqlalchemy==3.1.1
Flask-User==1.0.2.2
flask-wtf==1.2.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.6
MarkupSafe==2.1.5
mysqlclient==2.2.6
passlib==1.7.4
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
stripe==7.14.0
typing-extensions==4.12.2
urllib3==2.2.3
visitor==0.1.3
werkzeug==3.0.4
wtforms==3.1.2
zipp==3.20.2
